created the general consumable class. Right now in the test it is running with only pellets.
what still needs to be done is to extend the class to have individual classes for each consumable object in the game. This will be helpful because we can assign different point values for each consumable pretty easily, since most of the consumables don't have a ton of different functionality.